# packages/shared/openapi.yaml
openapi: 3.0.3
info:
  title: Gordon Ulen CPA Tax Helper API
  description: Tax preparation workflow automation system
  version: 1.0.0
  contact:
    name: RaveJedi Labs
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.taxhelper.com
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Auth JWT token
  schemas:
...
    DocumentType:
      type: string
      enum: [w2, 1099-r, 1099-g, 1099-int, 1099-div, 1099-nec, k1, other]
    DocumentStatus:
      type: string
      enum: [pending, processed, verified]
    Document:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        taxYear:
          type: integer
        type:
          $ref: '#/components/schemas/DocumentType'
        fileName:
          type: string
        fileUrl:
          type: string
        uploadedAt:
          type: string
          format: date-time
        uploadedBy:
          type: string
        ocrExtracted:
          type: boolean
          default: false
        extractedData:
          type: object
          additionalProperties: true
        status:
          $ref: '#/components/schemas/DocumentStatus'
    DocumentUpload:
      type: object
      required:
        - customerId
        - taxYear
        - fileName
      properties:
        customerId:
          type: string
        taxYear:
          type: integer
        type:
          $ref: '#/components/schemas/DocumentType'
        fileName:
          type: string

    # Routing Sheet
    RoutingSheet:
      type: object
      properties:
        dropOffDate:
          type: string
          format: date-time
        inPersonOrDropOff:
          type: string
          enum: [in-person, drop-off, portal]
        missingDocuments:
          type: array
          items:
            type: string
        notes:
          type: string

    # Payment
    PaymentStatus:
      type: string
      enum: [pending, partial, paid]
    PaymentMethod:
      type: string
      enum: [cash, check, card, square]
    Payment:
      type: object
      properties:
        amount:
          type: number
          format: float
        status:
          $ref: '#/components/schemas/PaymentStatus'
        method:
          $ref: '#/components/schemas/PaymentMethod'
        paidAt:
          type: string
          format: date-time

    # Tax Return
    ReturnType:
      type: string
      enum: ['1040', '1120', '1120s', '1065', '990']
    ReturnStatus:
      type: string
      enum:
        - intake
        - documents_pending
        - documents_complete
        - in_preparation
        - review_needed
        - waiting_on_client
        - ready_for_signing
        - completed
        - filed
        - picked_up
        - extension_needed
        - extension_filed
    TaxReturn:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        taxYear:
          type: integer
        returnType:
          $ref: '#/components/schemas/ReturnType'
        status:
          $ref: '#/components/schemas/ReturnStatus'
        assignedPreparer:
          type: string
        dueDate:
          type: string
          format: date-time
        extensionFiled:
          type: boolean
          default: false
        extensionDate:
          type: string
          format: date-time
        routingSheet:
          $ref: '#/components/schemas/RoutingSheet'
        payment:
          $ref: '#/components/schemas/Payment'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaxReturnCreate:
      type: object
      required:
        - customerId
        - taxYear
        - returnType
      properties:
        customerId:
          type: string
        taxYear:
          type: integer
        returnType:
          $ref: '#/components/schemas/ReturnType'
        assignedPreparer:
          type: string
        dueDate:
          type: string
          format: date-time
        routingSheet:
          $ref: '#/components/schemas/RoutingSheet'
    TaxReturnStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/ReturnStatus'
        notes:
          type: string

    # Appointment
...
    KanbanStatus:
      type: string
      enum: [backlog, in_progress, review, done]
    KanbanPriority:
      type: string
      enum: [low, medium, high]
    KanbanFeature:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        workflow:
          type: string
        status:
          $ref: '#/components/schemas/KanbanStatus'
        priority:
          $ref: '#/components/schemas/KanbanPriority'
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    KanbanFeatureUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        workflow:
          type: string
        status:
          $ref: '#/components/schemas/KanbanStatus'
        priority:
          $ref: '#/components/schemas/KanbanPriority'
        order:
          type: integer

    # User (Staff)
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # Pagination
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasMore:
          type: boolean

    # List responses
...
    DocumentList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    TaxReturnList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaxReturn'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    KanbanFeatureList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/KanbanFeature'
security:
  - BearerAuth: []
paths:
...
  /api/documents:
    get:
      summary: List documents
      operationId: listDocuments
      tags:
        - Documents
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
        - name: taxYear
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/DocumentType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DocumentStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentList'

    post:
      summary: Create document record
      operationId: createDocument
      tags:
        - Documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentUpload'
      responses:
        '201':
          description: Document record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
  /api/documents/upload:
    post:
      summary: Upload a document file
      operationId: uploadDocument
      tags:
        - Documents
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - customerId
                - taxYear
              properties:
                file:
                  type: string
                  format: binary
                customerId:
                  type: string
                taxYear:
                  type: integer
                type:
                  $ref: '#/components/schemas/DocumentType'
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
  /api/documents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get a document by ID
      operationId: getDocument
      tags:
        - Documents
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a document
      operationId: deleteDocument
      tags:
        - Documents
      responses:
        '204':
          description: Document deleted
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Tax Returns
  /api/returns:
    get:
      summary: List tax returns
      operationId: listReturns
      tags:
        - Tax Returns
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
        - name: taxYear
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReturnStatus'
        - name: assignedPreparer
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of tax returns
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxReturnList'

    post:
      summary: Create a tax return
      operationId: createReturn
      tags:
        - Tax Returns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaxReturnCreate'
      responses:
        '201':
          description: Tax return created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxReturn'
  /api/returns/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get a tax return by ID
      operationId: getReturn
      tags:
        - Tax Returns
      responses:
        '200':
          description: Tax return details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxReturn'
        '404':
          description: Tax return not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/returns/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    patch:
      summary: Update tax return status
      operationId: updateReturnStatus
      tags:
        - Tax Returns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaxReturnStatusUpdate'
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxReturn'
        '404':
          description: Tax return not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Appointments
...
  /api/kanban:
    get:
      summary: List kanban features
      operationId: listKanbanFeatures
      tags:
        - Kanban
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/KanbanStatus'
        - name: workflow
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of kanban features
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KanbanFeatureList'

    post:
      summary: Create a kanban feature
      operationId: createKanbanFeature
      tags:
        - Kanban
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                description:
                  type: string
                workflow:
                  type: string
                status:
                  $ref: '#/components/schemas/KanbanStatus'
                priority:
                  $ref: '#/components/schemas/KanbanPriority'
      responses:
        '201':
          description: Feature created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KanbanFeature'
  /api/kanban/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    patch:
      summary: Update a kanban feature
      operationId: updateKanbanFeature
      tags:
        - Kanban
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KanbanFeatureUpdate'
      responses:
        '200':
          description: Feature updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KanbanFeature'
        '404':
          description: Feature not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a kanban feature
      operationId: deleteKanbanFeature
      tags:
        - Kanban
      responses:
        '204':
          description: Feature deleted
        '404':
          description: Feature not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Users
...